+++ src/plugins/octetstream.js
@@ -4,7 +4,7 @@ import OctetStreamParser from '../parsers/OctetStream.js';
 
 export const octetStreamType = 'octet-stream';
 // the `options` is also available through the `options` / `formidable.options`
-export default function plugin(formidable, options) {
+export default async function plugin(formidable, options) {
   // the `this` context is always formidable, as the first argument of a plugin
   // but this allows us to customize/test each plugin
 
@@ -12,7 +12,7 @@ export default function plugin(formidable, options) {
   const self = this || formidable;
 
   if (/octet-stream/i.test(self.headers['content-type'])) {
-    init.call(self, self, options);
+    await init.call(self, self, options);
   }
   return self;
 }
@@ -20,7 +20,7 @@ export default function plugin(formidable, options) {
 // Note that it's a good practice (but it's up to you) to use the `this.options` instead
 // of the passed `options` (second) param, because when you decide
 // to test the plugin you can pass custom `this` context to it (and so `this.options`)
-function init(_self, _opts) {
+async function init(_self, _opts) {
   this.type = octetStreamType;
   const originalFilename = this.headers['x-file-name'];
   const mimetype = this.headers['content-type'];
@@ -31,7 +31,7 @@ function init(_self, _opts) {
   };
   const newFilename = this._getNewName(thisPart);
   const filepath = this._joinDirectoryName(newFilename);
-  const file = this._newFile({
+  const file = await this._newFile({
     newFilename,
     filepath,
     originalFilename,