+++ test/fixture/js/no-filename.js
@@ -14,7 +14,7 @@ const filename_name_http = [
     name: 'upload',
     originalFilename: 'plain.txt',
     fixture: 'filename-name',
-    sha1: 'b31d07bac24ac32734de88b3687dddb10e976872',
+    sha1: 'a47f7a8a7959f36c3f151ba8b0bd28f2d6b606e2',
   },
 ];